refactor(runtime): isolate startup internals
