+++ src/runtime/internal/lifecycle/session.go
@@ -1,4 +1,4 @@
-package runtime
+package lifecycle
 
 import (
 	"context"
@@ -6,15 +6,7 @@ import (
 	"sync"
 )
 
-type Session interface {
-	Ready() <-chan struct{}
-	Done() <-chan struct{}
-	Err() error
-	Stop()
-	Wait() error
-}
-
-type runningSession struct {
+type Session struct {
 	ctx     context.Context
 	readyCh <-chan struct{}
 	doneCh  chan struct{}
@@ -24,15 +16,15 @@ type runningSession struct {
 	err error
 }
 
-func newRunningSession(
+func New(
 	ctx context.Context,
 	readyCh <-chan struct{},
 	stop context.CancelFunc,
-) *runningSession {
+) *Session {
 	if ctx == nil {
 		ctx = context.Background()
 	}
-	return &runningSession{
+	return &Session{
 		ctx:     ctx,
 		readyCh: readyCh,
 		doneCh:  make(chan struct{}),
@@ -40,27 +32,40 @@ func newRunningSession(
 	}
 }
 
-func (s *runningSession) Ready() <-chan struct{} {
+func (s *Session) Ready() <-chan struct{} {
 	return s.readyCh
 }
 
-func (s *runningSession) Done() <-chan struct{} {
+func (s *Session) Done() <-chan struct{} {
 	return s.doneCh
 }
 
-func (s *runningSession) Err() error {
+func (s *Session) Err() error {
 	s.mu.Lock()
 	defer s.mu.Unlock()
 	return s.err
 }
 
-func (s *runningSession) Stop() {
+func (s *Session) Stop() {
 	s.stop()
 }
 
-func (s *runningSession) Wait() error {
+func (s *Session) Wait() error {
 	<-s.doneCh
 	return terminalErrOrNil(s.ctx, s.Err())
+}
+
+func (s *Session) Finish(err error) {
+	s.mu.Lock()
+	s.err = err
+	s.mu.Unlock()
+	close(s.doneCh)
+}
+
+func ClosedReadyCh() <-chan struct{} {
+	ch := make(chan struct{})
+	close(ch)
+	return ch
 }
 
 func terminalErrOrNil(ctx context.Context, err error) error {
@@ -70,11 +75,4 @@ func terminalErrOrNil(ctx context.Context, err error) error {
 		return err
 	}
 	return nil
-}
-
-func (s *runningSession) finish(err error) {
-	s.mu.Lock()
-	s.err = err
-	s.mu.Unlock()
-	close(s.doneCh)
 }
