+++ src/runtime/session.go
@@ -1,0 +1,9 @@
+package runtime
+
+type Session interface {
+	Ready() <-chan struct{}
+	Done() <-chan struct{}
+	Err() error
+	Stop()
+	Wait() error
+}
